--- posthog/ai/__init__.py
@@ -1,4 +1,0 @@
-from .providers.openai.openai import OpenAI
-from .providers.openai.openai_async import AsyncOpenAI
-
-__all__ = ["OpenAI", "AsyncOpenAI"]
