+++ posthog/ai/langchain/__init__.py
@@ -1,0 +1,3 @@
+from .callbacks import PosthogCallbackHandler
+
+__all__ = ["PosthogCallbackHandler"]
